qualtrics-api-project/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
